refactor: separate commander code from main functions
